domx: connection robustness improvements
CMA-enabled kernel for tuna devices will unload Ducati firmware when it's not
in use to free extra memory for applications.
On the first access to /dev/rpmsg-omx1 it will reload the firmware. However, in
the process of this reloading device drivers will be reset, therefore causing
errors when either opening the device or performing the first operation on it.
Moreover, the previously opened handle will be invalidated in this case.
Therefore, the connection loop is modified to retry both open() and first
ioctl()calls instead of just open() and to reopen the device on retries.

Also, because the reload can happen relatively fast when there are not too many
pages to migrate, we use smaller waiting period but increased number of retries
to improve camera start latency.

Change-Id: I4b919f457c0e3bc43f99381c405e364b3326f583